+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `StreamError::InvalidChar`.
+
 ### Changed
 - `Stream::consume_reference` will return only `InvalidReference` error from now.
 - `Error::InvalidTokenWithCause` merged into `Error::InvalidToken`.